requirements.txt
amqp==5.0.6
billiard==3.6.4.0
blinker==1.4
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dataclasses==0.6
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Mail==0.9.1
Flask-WTF==0.15.1
google-api-core==2.0.1
google-auth==2.0.2
google-cloud-secret-manager==2.7.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.40.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
libcst==0.3.20
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.0
prompt-toolkit==3.0.20
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.5.0
